requirements.txt
pandas==2.2.0
requests==2.31.0
plotly==5.18.0
kaleido==0.2.1
python-dotenv==1.0.0
